Move stored procedure to java implementation
Currently actions are created by stored procedure, this logic has
now been shifted into jpa repository logic.

We are making an assumption that there is a performance reason for
the stored proceedure, with social and census volumes in mind. To this
end we have maintained the insert being done by a select into rather than
the preferrable in object logic.

The integration tests are the same as for the previous commit,
guaranteeing that the behaviour is the same as the stored proceedure.