Change action rule offsets to static trigger dates
Action rules are currently triggered based of an offset number of days
from the scheduled actionplan start date. Instead of making this
calculation when creating, updating and triggering these action rules
it would be easier to have a static trigger date. This commit changes
action rules to work with static trigger dates.